+++ core-ng/src/main/java/core/framework/impl/cache/CacheManager.java
@@ -1,8 +1,9 @@
 package core.framework.impl.cache;
 
 import core.framework.cache.Cache;
+import core.framework.impl.reflect.GenericTypes;
+import core.framework.util.ASCII;
 import core.framework.util.Maps;
-import core.framework.util.Strings;
 
 import java.lang.reflect.Type;
 import java.time.Duration;
@@ -22,12 +23,13 @@ public class CacheManager {
         this.cacheStore = cacheStore;
     }
 
-    public <T> Cache<T> add(String name, Type valueType, Duration duration) {
+    public <T> Cache<T> add(Type valueType, Duration duration) {
         new CacheTypeValidator(valueType).validate();
 
+        String name = cacheName(valueType);
         CacheImpl<T> cache = new CacheImpl<>(name, valueType, duration, cacheStore);
         CacheImpl<?> previous = caches.putIfAbsent(name, cache);
-        if (previous != null) throw new Error(Strings.format("found duplicate cache name, name={}", name));
+        if (previous != null) throw new Error("found duplicate cache name, name=" + name);
         return cache;
     }
 
@@ -37,5 +39,15 @@ public class CacheManager {
 
     public List<CacheImpl<?>> caches() {
         return new ArrayList<>(caches.values());
+    }
+
+    // cache only supports List<T> and bean, may simplify further
+    String cacheName(Type valueType) {
+        if (valueType instanceof Class) {
+            return ASCII.toLowerCase(((Class<?>) valueType).getSimpleName());
+        } else if (GenericTypes.isGenericList(valueType)) {
+            return "list-" + ASCII.toLowerCase(GenericTypes.listValueClass(valueType).getSimpleName());
+        }
+        return ASCII.toLowerCase(valueType.getTypeName());
     }
 }