+++ core-ng/src/main/java/core/framework/impl/cache/CacheTypeValidator.java
@@ -14,7 +14,6 @@ class CacheTypeValidator {
         validator = new DataTypeValidator(instanceType);
         validator.allowChild = true;
         validator.allowTopLevelList = true;
-        validator.allowTopLevelValue = true;
     }
 
     public void validate() {